+++ example/src/ack-angular/AckApi.ts
@@ -5,38 +5,34 @@ import { Http, Response, Request } from '@angular/http';
 //Angular5
 //import { HttpClient, HttpRequest, HttpResponse, HttpEvent } from '@angular/common/http'
 
-import { AckCache } from './AckCache';
+import { AckCache, cacheModel } from './AckCache';
 import { AckQue } from './AckQue';
 
-export interface httpQueModel{
-  name          : string
-  maxTry?       : number
-  expires?      : number
-  maxAge?       : number
-  allowExpired? : boolean
-}
-
-export interface offlineMeta{
-  offlineId?   : number
+//specific to failed POST/PUT
+export interface sendFailMeta{
+  offlineId?   : number|string
   lastAttempt? : Date
   attempts?    : number
   maxTry?      : number
 }
 
 export interface httpOptions{
-  url          : string
-  method?      : 'GET'|'POST'|'PUT'|'PATCH'|'DELETE'|string
-  headers?     : any
-  timeout?     : number
-  queModel?    : httpQueModel|string
-  offlineMeta? : offlineMeta
-  promise?     : 'response'|'all'|'data'//typically just the body data is promised. Anything but data returns response
+  url             : string
+  params?         : any
+  body?           : any
+  method?         : 'GET'|'POST'|'PUT'|'PATCH'|'DELETE'|string
+  headers?        : any
+  timeout?        : number
+  offlineModel?   : cacheModel|string
+  sendFailMeta?   : sendFailMeta
+  promise?        : 'response'|'all'|'data'//typically just the body data is promised. Anything but data returns response
+  reportProgress? : boolean
 }
 
 export interface apiConfig{
   //promise? : 'all'|'data'//typically just the body data is promised
   baseUrl? : string
-  //params?  : object//query vars
+  //params?  : any//query vars
   $http?   : httpOptions
 }
 
@@ -57,8 +53,8 @@ TimeOutError.prototype = Object.create(Error.prototype)
   response:EventEmitter<Response> = new EventEmitter()
   AuthError:EventEmitter<Error> = new EventEmitter()
   ApiError:EventEmitter<Error> = new EventEmitter()
-  AckCache
-  AckQue
+  AckCache = new AckCache()
+  AckQue = new AckQue()
   config:apiConfig = <apiConfig>{
     baseUrl:'',
     $http:{
@@ -69,8 +65,6 @@ TimeOutError.prototype = Object.create(Error.prototype)
   }
 
   constructor(public HttpClient:Http){
-    this.AckCache = new AckCache()
-    this.AckQue = new AckQue()
     this.paramConfig()
   }
 
@@ -80,7 +74,7 @@ TimeOutError.prototype = Object.create(Error.prototype)
   /** START HERE. Handlers must be registered before sending requests
     @options{maxTry:50}
   */
-  registerHandler(name:string, handler?:(config:httpOptions)=>any, options?:offlineMeta){
+  registerHandler(name:string, handler?:(config:httpOptions)=>any, options?:sendFailMeta){
     options = options || {maxTry:50}
     handler = handler || (config=>this.request(config))
     this.AckQue.registerHandler(name, handler)
@@ -113,11 +107,11 @@ TimeOutError.prototype = Object.create(Error.prototype)
   }
 
   /** method all request transactions tunnel thru to instead try for cache first
-    HINT: @config.queModel, when defined:
+    HINT: @config.offlineModel, when defined:
       - POST/PUT/PATCH requests goto que if they fail.
       - GET responses are cached with optional expires or maxAge option
   */
-  request(config:httpOptions){
+  request(config:httpOptions):Promise<Response>{
     const defaults:httpOptions = { ...this.config.$http }
 
     defaults.headers = { ...this.config.$http.headers }//cause a deeper clone and break memory ref with apiConfig.$http
@@ -129,50 +123,69 @@ TimeOutError.prototype = Object.create(Error.prototype)
     //Object.assign(request.headers, this.config.$http.headers)//enforced config/defaults . apiConfig act as an overwriter of headers
 
     //has cache instructions?
-    if( request.queModel ){
-      return this.requestQueModel(request)
+    if( request.offlineModel ){
+      return this.requestOfflineModel(request)
     }
 
     return this._fetch(request)
   }
 
   getCacheByNamedRequest(request:httpOptions){
-    const queModel = <httpQueModel>request.queModel
-    return this.AckCache.get( queModel.name )
+    const offlineModel = <cacheModel>request.offlineModel
+    return this.AckCache.get( offlineModel.name )
     .then(routes=>{
       routes = routes || {}
-      return routes[request.url]
+      const cacheName = this.getSotageNameByRequest(request)
+console.log('cacheName', cacheName, routes[cacheName], routes)
+      return routes[ cacheName ]//get cache by url
     })
     .then(cache=>this.processCacheGet(cache,request))
   }
 
-  requestQueModel(request:httpOptions){
-    let queModel = <httpQueModel>request.queModel
-    if(queModel && queModel.constructor==String){
-      request.queModel = queModel = <httpQueModel>{ name:request.queModel }
+  getSotageNameByRequest(request:httpOptions):string{
+    if( request.params ){
+      const paramKeys = Object.keys(request.params)
+
+      if(!paramKeys.length)return request.url
+      
+      let url = request.url + (request.url.search(/\?/)>=0 ? '&' : '?')
+      paramKeys.sort().forEach(name=>url+=name+'='+request.params[name]+'&')
+
+      return url.substring(0, url.length-1)//remove last amp
+    }
+
+    return request.url
+  }
+
+  requestOfflineModel(request:httpOptions):Promise<Response>{
+    let offlineModel = <cacheModel>request.offlineModel
+    
+    if(offlineModel && offlineModel.constructor==String){
+      request.offlineModel = offlineModel = <cacheModel>{ name:request.offlineModel }
     }
 
     if (request.method === "GET") {
       return this.getCacheByNamedRequest(request)
     }
 
-    this.AckQue.paramHandler(queModel.name,config=>this._fetch(config))
+    this.AckQue.paramHandler(offlineModel.name,config=>this._fetch(config))
 
     //request is a PUT, POST, PATCH, or DELETE
     return this._fetch(request)
     .catch( e=>this.postRequestFail(e,request) )//if fail, save config for later
   }
 
-  processCacheGet(cache, cfg:httpOptions){
+  processCacheGet(cache:any, cfg:httpOptions):Promise<Response>{
     if(cache==null)return this._fetch(cfg)
 
-    const queModel = <httpQueModel>cfg.queModel
-    return this.AckCache.cacheToReturn(queModel.name, cache, queModel)
+    const offlineModel = <cacheModel>cfg.offlineModel
+    
+    return this.AckCache.cacheToReturn(offlineModel.name, cache, offlineModel)
     .then(rtn=>{
-      const willExpire = rtn && this.AckCache.optionsKillCache(cfg.queModel)
+      const willExpire = rtn && this.AckCache.optionsKillCache( offlineModel )
 
       if( !willExpire ){
-        console.log('AckApi fetched cache that will never expire. Set queModel.expires=0 or queModel.maxAge=0 to avoid this message')
+        console.log('AckApi fetched cache that will never expire. Set offlineModel.expires=0 or offlineModel.maxAge=0 to avoid this message')
       }
 
       if( rtn!=null ){
@@ -188,20 +201,20 @@ TimeOutError.prototype = Object.create(Error.prototype)
 
     if(!saveWorthy)return Promise.reject(e)
 
-    request.offlineMeta = request.offlineMeta || <offlineMeta>{}
-    request.offlineMeta.offlineId = Date.now()
-    request.offlineMeta.lastAttempt = new Date()
-    request.offlineMeta.attempts = request.offlineMeta.attempts==null ? 1 : ++request.offlineMeta.attempts
-    request.offlineMeta.maxTry = request.offlineMeta.maxTry || 50
-    const tryAgainLater = request.offlineMeta.maxTry && request.offlineMeta.attempts <= request.offlineMeta.maxTry
+    request.sendFailMeta = request.sendFailMeta || <sendFailMeta>{}
+    request.sendFailMeta.offlineId = Date.now()//this.getSotageNameByRequest(request)
+    request.sendFailMeta.lastAttempt = new Date()
+    request.sendFailMeta.attempts = request.sendFailMeta.attempts==null ? 1 : ++request.sendFailMeta.attempts
+    request.sendFailMeta.maxTry = request.sendFailMeta.maxTry || 50
+    const tryAgainLater = request.sendFailMeta.maxTry && request.sendFailMeta.attempts <= request.sendFailMeta.maxTry
 
-    e.offlineMeta = request.offlineMeta
+    e.sendFailMeta = request.sendFailMeta
 
     if(tryAgainLater){
       const requestSave = Object.assign({}, request)
-      delete requestSave.queModel//only used for GET method
-      const queModel = <httpQueModel>request.queModel
-      return this.AckQue.set(queModel.name, requestSave)
+      delete requestSave.offlineModel//only used for GET method
+      const offlineModel = <cacheModel>request.offlineModel
+      return this.AckQue.set(offlineModel.name, requestSave)
       .then( ()=>Promise.reject(e) )
     }
 
@@ -274,7 +287,7 @@ TimeOutError.prototype = Object.create(Error.prototype)
     const isDataMode = !request.promise || request.promise=='data'
     const output = isDataMode ? (response['data']||data) : response
 
-    if (request.method === "GET" && request.queModel) {
+    if (request.method === "GET" && request.offlineModel) {
       return this.requestResponseToCache(request, output)
       .then( ()=>output )
     }
@@ -303,31 +316,26 @@ TimeOutError.prototype = Object.create(Error.prototype)
   }
 
   requestResponseToCache(request, output){
-    const cachename = request.queModel.name || request.queModel
+    const cachename = request.offlineModel.name || request.offlineModel
     return this.AckCache.get(cachename)
     .then(routes=>{
       routes = routes || {}
-      routes[ request.url ] = {cache:output}
-      this.AckCache.dataOptionsCache(routes[ request.url ], request.queModel, output)
+      const cacheName = this.getSotageNameByRequest(request)
+      routes[ cacheName ] = <cacheModel>{cache:output}
+      this.AckCache.dataOptionsCache(routes[ request.url ], request.offlineModel, output)
       return routes
     })
     .then(routes=>this.AckCache.set(cachename, routes))
   }
   
-  /**
-    @path:url
-    @config:{
-      params:{}//url parameters
-    }
-  */
-  get(path, config?) {
+  get(path:string, config?:httpOptions) {
     const cfg = Object.assign({}, config)
     cfg.method = "GET"
     cfg.url = path
     return this.request( cfg )
   }
 
-  post(path, data, config?) {
+  post(path:string, data:any, config?:httpOptions) {
     const cfg = Object.assign({}, config)
     cfg.method = "POST"
     //cfg.data = data
@@ -336,7 +344,7 @@ TimeOutError.prototype = Object.create(Error.prototype)
     return this.request( cfg )
   }
 
-  delete(path, config?) {
+  delete(path:string, config?:httpOptions) {
     const cfg = Object.assign({}, config)
     cfg.method = "DELETE"
     cfg.url = path
@@ -363,24 +371,25 @@ TimeOutError.prototype = Object.create(Error.prototype)
 }
 
 /** prevent angular1 from assuming the header to send is application/json */
-function upgradeConfig(cfg){
+function upgradeConfig(cfg:httpOptions){
   cfg.method = cfg.method || 'GET'
 
   cfg.reportProgress = cfg.reportProgress || false
 
-  const isFormData = cfg.data && FormData && cfg.data.constructor==FormData
+  //A trick from AngularJs to prevent header auto population
+  /*const isFormData = cfg.body && FormData && cfg.body.constructor==FormData
   if(isFormData){
     const preventAutoContentType =  !cfg.headers || Object.keys(cfg.headers).filter(h=>h.search(/content-type/i)<0)
     
     if(preventAutoContentType){
       cfg.headers['Content-Type'] = undefined//'multipart/form-data;'
     }
-  }
+  }*/
 
   //cast url variables nulls to empty-string
-  if(cfg.params){
+  if( cfg.params ){
     for(let key in cfg.params){
-      if(cfg.params[key]==null){
+      if( cfg.params[key]==null ){
         cfg.params[key]=''
       }
     }