+++ example/src/ack-angular/AckOffline.ts
@@ -7,23 +7,23 @@ import { Injectable } from '@angular/core';
   public prefix:string="offline"
 
   /** save to browser offline memory with datetime stamps offlineSavedAt and offlineCreatedAt */
-  set(name, data) {
+  set(name:string, data:any) {
     if(data && data.constructor==Object)data.offlineSavedAt = new Date().toString()
     if(data && data.constructor==Object)data.offlineCreatedAt = data.offlineCreatedAt || new Date().toString()
     
     return localForage.setItem(this.prefix+'-'+name, data)
   }
 
-  get(name) {
+  get(name:string) {
     return localForage.getItem(this.prefix+'-'+name)
   }
 
-  remove(name){
+  remove(name:string){
     return localForage.removeItem(this.prefix+'-'+name)
   }
 
   /** aka of remove */
-  clear(name){return this.remove(name)}
+  clear(name:string){return this.remove(name)}
 
   clearAll() {
     return this.promiseNameArray().then((keys) => keys.forEach(name=>this.remove(name)))