+++ src/AckCache.ts
@@ -1,48 +1,57 @@
 import { Injectable } from '@angular/core';
 import { AckOffline } from './AckOffline';
 
+export interface cacheModel{
+  name?         : string
+  maxTry?       : number
+  expires?      : number
+  maxAge?       : number
+  allowExpired? : boolean
+  param?        : object//when cache not defined, use this
+}
+
+
 /** Cache logic that provides timestamping and timingout using AckOffline as storage */
 @Injectable() export class AckCache extends AckOffline{
-  public prefix:string="offline-cache"
+  prefix:string="offline-cache"
   
   validate(data, config) {
     const exists = data !== null && typeof data.cache !== "undefined"
     return exists && !this.hasExpired(data._timestamp, data.expires||config.expires)
   }
 
-  optionsKillCache(options){
+  optionsKillCache(options:cacheModel){
     return options.maxAge==null && options.expires==null ? false : true
   }
 
-  hasMaxAged(stamp, maxAge){
+  hasMaxAged(stamp:number, maxAge:number){
     const diff = Date.now() - maxAge
     const expired = stamp<=diff
     return maxAge && expired ? true : false
   }
 
-  hasExpired(stamp, expires){
+  hasExpired(stamp:number, expires:number){
     const expired = Date.now() >= new Date(expires).getTime()
     return expires && expired ? true : false
   }
 
-  param(name, options){
-    options = options || {}
-    options.param = options.param || []
+  param(name:string, options:cacheModel=<cacheModel>{}){
+    options.param = options.param || []//good idea? Assumption data returned is an array
     return this.getCache(name, options)
   }
 
   /** aka param */
-  paramCache(name, options){
+  paramCache(name:string, options:cacheModel){
     return this.param(name, options)    
   }
 
-  paramSave(name, options){
+  paramSave(name:string, options:cacheModel){
     return this.paramCache(name, options)
     .then(items=>this.setCache(name,items))
   }
 
   /** aka paramSave */
-  paramSaveCache(name, options){
+  paramSaveCache(name:string, options:cacheModel){
     return this.paramSave(name, options)
   }
 
@@ -52,20 +61,20 @@ import { AckOffline } from './AckOffline';
       maxAge:number - 
     }
   */
-  cacheToReturn(name, data, options){
-    if(data['cache']==null && data['_timestamp']==null){
+  cacheToReturn(name:string, data:any, options:cacheModel):any{
+    if(data.cache==null && data._timestamp==null){
       return Promise.resolve( data )//bad data catch
     }
 
-    const expires = data['expires'] || options.expires
-    const maxAge = data['maxAge'] || options.maxAge
+    const expires = data.expires || options.expires
+    const maxAge = data.maxAge || options.maxAge
     
     if(options.allowExpired || (!expires && !maxAge)){
-      return Promise.resolve( data['cache'] )
+      return Promise.resolve( data.cache )
     }
 
-    const expired = expires && this.hasExpired(data['_timestamp'], expires)
-    const isMaxed = maxAge && this.hasMaxAged(data['_timestamp'], maxAge)
+    const expired = expires && this.hasExpired(data._timestamp, expires)
+    const isMaxed = maxAge && this.hasMaxAged(data._timestamp, maxAge)
 
     if(expired || isMaxed){
       return this.selfDestructData(name,data)
@@ -92,9 +101,7 @@ import { AckOffline } from './AckOffline';
     return isMemMaxed || isMemExpired
   }
 
-  get(name, options?){
-    options = options || {expires:null}
-    
+  get(name:string, options:cacheModel=<cacheModel>{expires:null}){
     return super.get(name)
     .then(data=>{
       if(data){
@@ -108,42 +115,41 @@ import { AckOffline } from './AckOffline';
   }
 
   /** aka get */
-  getCache(name, options){
+  getCache(name:string, options?:cacheModel){
     return this.get(name, options)
   }
 
   /** paste cache over cache, leave all else alone */
-  dataOptionsCache(allCache, options, cache){
-    options = options || {}
-    const newOptions = {}
+  dataOptionsCache(allCache, options:cacheModel=<cacheModel>{}, cache:any){
+    const newOptions:any = {}
     
-    newOptions['_timestamp'] = Date.now()
-    if(options.expires)newOptions['expires'] = options.expires
-    if(options.maxAge)newOptions['maxAge'] = options.maxAge
+    newOptions._timestamp = Date.now()
+    if(options.expires)newOptions.expires = options.expires
+    if(options.maxAge)newOptions.maxAge = options.maxAge
     
     allCache = allCache && allCache.constructor!=String ? allCache : {}
 
     Object.assign(allCache, newOptions)
     if(cache && cache.constructor==String){
-      allCache['cache'] = cache
-    }else if(allCache['cache'] && allCache['cache'].constructor!=String){
-      Object.assign(allCache['cache'], cache)      
+      allCache.cache = cache
+    }else if(allCache.cache && allCache.cache.constructor!=String){
+      Object.assign(allCache.cache, cache)      
     }else{
-      allCache['cache'] = cache
+      allCache.cache = cache
     }
 
     return allCache
   }
 
-  set(name, cache, options?) {
-    options = options || {}
+  set(name:string, cache:any, options:cacheModel=<cacheModel>{}) {
+    options.name = options.name || name
     return super.get(name)
     .then( allCache=>this.dataOptionsCache(allCache, options, cache) )
     .then( data=>super.set(name, data) )
   }
 
   /** aka set */
-  setCache(name, cache, options?){
+  setCache(name:string, cache:any, options?:cacheModel){
     return this.set(name, options)
   }
 }